Renamed AssetTestUtils and ImageAssetTests to stick to more descriptive naming conventions
