+++ game/tests/src/com/eng1/gdxtesting/AssetTests/SkinAssetUnitTests.java
@@ -5,7 +5,7 @@ import com.eng1.gdxtesting.GdxTestRunner;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import static com.eng1.gdxtesting.AssetTests.AssetTestsUtils.assetTestReflectionStaticPath;
+import static com.eng1.gdxtesting.AssetTests.AssetUnitTestsUtils.assetTestReflectionStaticPath;
 
 @RunWith(GdxTestRunner.class)
 public class SkinAssetUnitTests {
